feat: stripe payment

+++ src/services/wallet.service.ts
@@ -5,10 +5,11 @@ import { isValidObjectId } from "mongoose";
 
 import { BCRYPT_SALT } from "../config";
 import MailService from "./mail.service";
+import StripeUtil from "../utils/stripe";
 import Token from "../models/token.model";
 import Wallet from "../models/wallet.model";
-import CustomError from "../utils/graphql/custom-error";
 import useTransaction from "../utils/use-transaction";
+import CustomError from "../utils/graphql/custom-error";
 
 import type { ClientSession } from "mongoose";
 
@@ -24,6 +25,15 @@ class WalletService {
         if (!isValidObjectId(userId)) throw new CustomError("Invalid userId");
 
         const wallet = await Wallet.findOne({ userId });
+        if (!wallet) throw new CustomError("Wallet not found");
+
+        return wallet;
+    }
+
+    async getOne(walletId: string) {
+        if (!isValidObjectId(walletId)) throw new CustomError("Invalid walletId");
+
+        const wallet = await Wallet.findOne({ _id: walletId });
         if (!wallet) throw new CustomError("Wallet not found");
 
         return wallet;
@@ -66,11 +76,92 @@ class WalletService {
         };
     }
 
+    async initializeDeposit(userId: string, amount: number, currencyCode: string) {
+        const { default: UserService } = await import("./user.service");
+        const { default: CurrencyService } = await import("./currency.service");
+        const { default: NotificationService } = await import("./notification.service");
+
+        const user = await UserService.getOne(userId);
+        const wallet = await this.getByUserId(userId);
+
+        if (!amount) throw new CustomError("amount is required");
+        if (!Number.isInteger(amount)) throw new CustomError("Amount must be an integer");
+        if (amount < 10) throw new CustomError("Minimum amount is 10");
+        if (!currencyCode) throw new CustomError("currency is required");
+
+        const currency = await CurrencyService.getByCode(currencyCode);
+
+        const purchaseData = {
+            amount,
+            walletId: wallet.id,
+            currency: currency.code,
+            price: amount * currency.gildRate * (currency.isZeroDecimal ? 1 : 100),
+            user: {
+                email: user.email,
+                name: `${user.name}`
+            }
+        };
+
+        const intent = await StripeUtil.purchaseGild(purchaseData);
+
+        await NotificationService.create({
+            sourceId: "system",
+            receiverId: user.id,
+            title: "Deposit Initiated",
+            message: `You have initiated a deposit of ${amount} GILD to your Gild wallet`
+        });
+
+        return intent.client_secret as string;
+    }
+
+    async completeDeposit(metadata: DepositMetadata, stripePaymentId: string) {
+        const { default: TransactionService } = await import("./transaction.service");
+        const { default: NotificationService } = await import("./notification.service");
+
+        const price = parseInt(metadata.price);
+        const amount = parseInt(metadata.amount);
+
+        const wallet = await this.getOne(metadata.wallet_id);
+
+        await useTransaction(async (session: ClientSession) => {
+            await Wallet.findOneAndUpdate({ _id: metadata.wallet_id }, { $inc: { balance: amount } }, { session });
+
+            await NotificationService.create(
+                {
+                    sourceId: "system",
+                    receiverId: wallet.userId,
+                    title: "Deposit Completed",
+                    message: `You have successfully deposited ${amount} GILD to your Gild wallet`
+                },
+                session
+            );
+
+            await TransactionService.recordDeposit(
+                { walletId: metadata.wallet_id, userId: wallet.userId, currency: metadata.currency, amount, price, stripePaymentId },
+                session
+            );
+        });
+    }
+
     async initializeTransfer(senderId: string, receiverId: string, amount: number) {
         const { default: UserService } = await import("./user.service");
+        const { default: TransactionService } = await import("./transaction.service");
         const { default: NotificationService } = await import("./notification.service");
 
         const sender = await UserService.getOne(senderId);
+        const senderWallet = await this.getByUserId(senderId);
+
+        if (senderId === receiverId) throw new CustomError("invalid transfer");
+        if (senderWallet.balance < amount) throw new CustomError("insufficient funds");
+
+        const last24Hours = await TransactionService.transferInLast24Hours(senderId);
+        if (last24Hours.count >= 3) throw new CustomError("you have reached the maximum number of transfers per day");
+        if (last24Hours.totalAmount + amount > 1000) throw new CustomError("you have reached the maximum amount of transfers per day");
+
+        if (!amount) throw new CustomError("amount is required");
+        if (!Number.isInteger(amount)) throw new CustomError("amount must be an integer");
+        if (amount < 10) throw new CustomError("minimum transfer is 10 Gild tokens per transaction");
+        if (amount > 250) throw new CustomError("maximum transfer is 250 Gild tokens per transaction");
 
         const token = await Token.findOne({ userId: senderId, type: "transfer_gild" });
         if (token) await token.deleteOne();
@@ -88,8 +179,7 @@ class WalletService {
             gildTransfer: { amount, receiverId }
         }).save();
 
-        // await new MailService(sender).sendTransferOTP(OTP);
-        console.log(`OTP: ${OTP}`);
+        await new MailService(sender).sendTransferOTP(OTP);
 
         await NotificationService.create({
             title: "Hello world",
@@ -107,13 +197,9 @@ class WalletService {
         const senderWallet = await this.getByUserId(senderId);
         const receiverWallet = await this.getByUserId(data.receiverId);
 
-        if (senderId === data.receiverId) throw new CustomError("invalid action");
-
         if (!data.OTP) throw new CustomError("OTP is required");
         if (!data.amount) throw new CustomError("amount is required");
-        if (!Number.isInteger(data.amount)) throw new CustomError("amount must be an integer");
-        if (data.amount < 10) throw new CustomError("minimum transfer is 10 Gild tokens per transaction");
-        if (data.amount > 250) throw new CustomError("maximum transfer is 250 Gild tokens per transaction");
+        if (!data.receiverId) throw new CustomError("receiverId is required");
 
         const token = await Token.findOne({
             userId: senderId,
@@ -138,6 +224,91 @@ class WalletService {
 
             await token.deleteOne({ session });
             await TransactionService.recordTransfer({ senderId, receiverId: data.receiverId, amount: data.amount }, session);
+        });
+
+        return true;
+    }
+
+    async initializeWithdrawal(userId: string, amount: number) {
+        const { default: UserService } = await import("./user.service");
+        const { default: NotificationService } = await import("./notification.service");
+
+        if (!amount) throw new CustomError("amount is required");
+        if (!Number.isInteger(amount)) throw new CustomError("Amount must be an integer");
+        if (amount < 50) throw new CustomError("Minimum amount is 50");
+
+        const wallet = await this.getByUserId(userId);
+        if (wallet.balance < amount) throw new CustomError("insufficient funds");
+
+        const user = await UserService.getOne(userId);
+        if (!user.stripeAccountId) throw new CustomError("you have not setup a stripe connect account");
+
+        // Delete any existing withdrawal token
+        const token = await Token.findOne({ userId, type: "withdraw_gild" });
+        if (token) await token.deleteOne();
+
+        const nanoidOTP = customAlphabet("012345789", 6);
+        const OTP = nanoidOTP();
+        const hashedOTP = await bcrypt.hash(OTP, BCRYPT_SALT);
+
+        await new Token({
+            userId,
+            token: hashedOTP,
+            type: "withdraw_gild",
+            expiresAt: Date.now() + ms("15m"),
+            gildWithdrawal: { amount, walletId: wallet.id }
+        }).save();
+
+        await new MailService(user).sendTransferOTP(OTP);
+
+        await NotificationService.create({
+            sourceId: "system",
+            receiverId: user.id,
+            title: "Withdrawal Initiated",
+            message: `You have initiated a withdrawal of ${amount} GILD from your Gild wallet`
+        });
+
+        return true;
+    }
+
+    async completeWithdrawal(userId: string, data: GildWithdrawalInput) {
+        const { default: UserService } = await import("./user.service");
+        const { default: TransactionService } = await import("./transaction.service");
+        const { default: NotificationService } = await import("./notification.service");
+
+        if (!data.OTP) throw new CustomError("OTP is required");
+        if (!data.amount) throw new CustomError("amount is required");
+
+        const wallet = await this.getByUserId(userId);
+        const user = await UserService.getOne(userId);
+
+        const token = await Token.findOne({
+            userId,
+            type: "withdraw_gild",
+            "gildWithdrawal.amount": data.amount,
+            "gildWithdrawal.walletId": wallet.id
+        });
+        if (!token) throw new CustomError("withdrawal cannot be completed");
+
+        const isOTPValid = await bcrypt.compare(data.OTP, token.token);
+        if (!isOTPValid) throw new CustomError("invalid OTP");
+
+        if (wallet.balance < data.amount) throw new CustomError("insufficient funds");
+
+        await useTransaction(async (session: ClientSession) => {
+            await Wallet.findOneAndUpdate({ _id: wallet.id }, { balance: wallet.balance - data.amount }, { session });
+
+            await NotificationService.create({
+                sourceId: "system",
+                receiverId: user.id,
+                title: "Withdrawal Completed",
+                message: `You have successfully withdrawn ${data.amount} GILD from your Gild wallet`
+            }, session);
+            
+            await token.deleteOne({ session });
+            await TransactionService.recordWithdrawal({ userId, amount: data.amount, walletId: wallet.id }, session);
+
+            await StripeUtil.payout(user.stripeAccountId, data.amount, "Gild Wallet Withdrawal");
         });
 
         return true;
